+++ apps/astarte_appengine_api/CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 
 ## [0.10.0-beta.2] - Unreleased
+### Added
+- Added value type and size validation.
+
 ### Changed
 - "data" key is used instead of "value" when PUT/POSTing a value to an interface.